refactor: organize files in packages
- use MVC pattern as reference to organize files
