+++ src/main/kotlin/lotto/Application.kt
@@ -1,5 +1,10 @@
 package lotto
 
+import lotto.controller.LottoApp
+import lotto.model.LottoService
+import lotto.view.InputView
+import lotto.view.OutputView
+
 fun main() {
     val lottoService = LottoService()
     val inputView = InputView()
