+++ src/main/kotlin/lotto/model/LottoGenerator.kt
@@ -1,6 +1,8 @@
-package lotto
+package lotto.model
 
 import camp.nextstep.edu.missionutils.Randoms
+import lotto.Lotto
+import lotto.util.LottoConstants
 
 class LottoGenerator {
