+++ src/main/kotlin/lotto/util/LottoConstants.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 object LottoConstants {
     const val TICKET_PRICE = 1000
