+++ src/main/kotlin/lotto/util/Validator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 object Validator {
