+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,6 +1,8 @@
-package lotto
+package lotto.view
 
 import camp.nextstep.edu.missionutils.Console
+import lotto.util.Messages
+import lotto.util.Validator
 
 class InputView {
