+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -1,4 +1,8 @@
-package lotto
+package lotto.view
+
+import lotto.Lotto
+import lotto.util.Messages
+import lotto.model.Rank
 
 class OutputView {
